Moved App and created base folders
